// app/utils/sanity/types/sanity.types.ts
export type UnfurledUrl = {
  _type: 'unfurledUrl'
  url?: string
}

export type Youtube = {
  _type: 'youtube'
  url?: string
}

export type Twitter = {
  _type: 'twitter'
  url?: string
}

export type PortableText = Array<
  | {
      children?: Array<{
        marks?: Array<string>
        text?: string
        _type: 'span'
        _key: string
      }>
      style?: 'normal' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote'
      listItem?: 'bullet' | 'number'
      markDefs?: Array<{
        href?: string
        _type: 'link'
        _key: string
      }>
      level?: number
      _type: 'block'
      _key: string
    }
  | {
      asset?: {
        _ref: string
        _type: 'reference'
        _weak?: boolean
        [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
      }
      hotspot?: SanityImageHotspot
      crop?: SanityImageCrop
      src?: string
      caption?: string
      maxWidth?: number
      alt?: string
      _type: 'imageWithMetadata'
      _key: string
    }
  | ({
      _key: string
    } & Code)
  | ({
      _key: string
    } & CodeSandbox)
  | ({
      _key: string
    } & CodePen)
  | ({
      _key: string
    } & Youtube)
  | ({
      _key: string
    } & Twitter)
  | ({
      _key: string
    } & Iframe)
  | ({
      _key: string
    } & InfoBlock)
  | ({
      _key: string
    } & UnfurledUrl)
>

export type ImageWithMetadata = {
  _type: 'imageWithMetadata'
  asset?: {
    _ref: string
    _type: 'reference'
    _weak?: boolean
    [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
  }
  hotspot?: SanityImageHotspot
  crop?: SanityImageCrop
  src?: string
  caption?: string
  maxWidth?: number
  alt?: string
}

export type DescriptionText = Array<{
  children?: Array<{
    marks?: Array<string>
    text?: string
    _type: 'span'
    _key: string
  }>
  style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote'
  listItem?: 'bullet' | 'number'
  markDefs?: Array<{
    href?: string
    _type: 'link'
    _key: string
  }>
  level?: number
  _type: 'block'
  _key: string
}>

export type CodeSandbox = {
  _type: 'codeSandbox'
  url?: string
}

export type CodePen = {
  _type: 'codePen'
  url?: string
}

export type InfoBlock = {
  _type: 'infoBlock'
  content?: Array<{
    children?: Array<{
      marks?: Array<string>
      text?: string
      _type: 'span'
      _key: string
    }>
    style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote'
    listItem?: 'bullet' | 'number'
    markDefs?: Array<{
      href?: string
      _type: 'link'
      _key: string
    }>
    level?: number
    _type: 'block'
    _key: string
  }>
}

export type Iframe = {
  _type: 'iframe'
  src?: string
  height?: number
}
// ...
export type Post = {
  _id: string
  _type: 'post'
  _createdAt: string
  _updatedAt: string
  _rev: string
  type?: 'article' | 'video' | 'podcast'
  language?: 'en-US' | 'nb-NO' | 'nn-NO'
  embedUrl?: string
  podcastLength?: number
  title?: string
  slug?: Slug
  canonicalUrl?: string
  description?: DescriptionText
  authors?: Array<{
    _ref: string
    _type: 'reference'
    _weak?: boolean
    _key: string
    [internalGroqTypeReferenceTo]?: 'author'
  }>
  coverImage?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    src?: string
    hideFromPost?: boolean
    _type: 'image'
  }
  availableFrom?: string
  tags?: Array<{
    _ref: string
    _type: 'reference'
    _weak?: boolean
    _key: string
    [internalGroqTypeReferenceTo]?: 'tag'
  }>
  keywords?: Array<string>
  content?: PortableText
  relatedLinks?: Array<{
    title?: string
    description?: string
    url?: string
    _type: 'relatedLink'
    _key: string
  }>
  priority?: number
}

export type SanityImageCrop = {
  _type: 'sanity.imageCrop'
  top?: number
  bottom?: number
  left?: number
  right?: number
}

export type SanityImageHotspot = {
  _type: 'sanity.imageHotspot'
  x?: number
  y?: number
  height?: number
  width?: number
}
// ...
export type Code = {
  _type: 'code'
  language?: string
  filename?: string
  code?: string
  highlightedLines?: Array<number>
}

export type MediaTag = {
  _id: string
  _type: 'media.tag'
  _createdAt: string
  _updatedAt: string
  _rev: string
  name?: Slug
}

export type Slug = {
  _type: 'slug'
  current?: string
  source?: string
}
// ...
export declare const internalGroqTypeReferenceTo: unique symbol